% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mosdepth.R
\name{mosdepth_calculate_proportion_of_genome_with_depth_over_threshold}
\alias{mosdepth_calculate_proportion_of_genome_with_depth_over_threshold}
\title{Calculate proportion of genome covered at given depth}
\usage{
mosdepth_calculate_proportion_of_genome_with_depth_over_threshold(
  mosdepth_outputs,
  depth_threshold = 10,
  chr = "total"
)
}
\arguments{
\item{mosdepth_outputs}{result of running \code{\link[=mosdepth_collect_output_files]{mosdepth_collect_output_files()}}}

\item{depth_threshold}{depth threshold}

\item{chr}{limit analysis to specific chromosome. The default of 'total' will calculated  \strong{'proportion of all bases covered at depth' >= depth_threshold} across all chromosomes in the file}
}
\value{
'proportion of all bases covered at depth' >= depth_threshold (numeric)
}
\description{
Calculate proportion of genome covered at given depth
}
\examples{
mosdepth_dir = system.file("mosdepth_test_files",package="mosdepth")

mosdepth_collect_output_files(prefix = "E_coli", mosdepth_dir) |>
  mosdepth_calculate_proportion_of_genome_with_depth_over_threshold(depth_threshold = 100)
}
