% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectingMosdepthOutput.R
\name{mosdepth_collect_output_files}
\alias{mosdepth_collect_output_files}
\title{Collect mosdepth files}
\usage{
mosdepth_collect_output_files(prefix, directory_path = ".")
}
\arguments{
\item{prefix}{prefix supplied in original mosdepth call (string)}

\item{directory_path}{directory containing the mosdepth files (string)}
}
\value{
an s4 object of class 'MosdepthOutputFiles'. This can be fed into many other functions to analyse/visualise the results
}
\description{
Creates an object that keeps track of the location of the mosdepth files
}
\examples{
mosdepth_dir = system.file("mosdepth_test_files", package="mosdepth")

mosdepth_collect_output_files(
  prefix="E_coli",
  directory_path = mosdepth_dir
)


}
