% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_mosdepth.R
\name{mosdepth_plot_cumulative_coverage_distribution}
\alias{mosdepth_plot_cumulative_coverage_distribution}
\title{Plot Cumulative Coverage Distribution}
\usage{
mosdepth_plot_cumulative_coverage_distribution(
  mosdepth_outputs,
  chromosomes_of_interest = NULL
)
}
\arguments{
\item{mosdepth_outputs}{result of running \code{\link[=mosdepth_collect_output_files]{mosdepth_collect_output_files()}}}

\item{chromosomes_of_interest}{a subset of chromosomes to display. May include contain  'total' (character)}
}
\value{
ggplot object
}
\description{
Plot Cumulative Coverage Distribution
}
\examples{
mosdepth_dir = system.file("mosdepth_test_files",package="mosdepth")

# Plot Cumulative Coverage
mosdepth_collect_output_files(prefix = "E_coli", mosdepth_dir) |>
  mosdepth_plot_cumulative_coverage_distribution()
}
