% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectingMosdepthOutput.R
\name{mosdepth_retrieve_file}
\alias{mosdepth_retrieve_file}
\title{Retrive Mosdepth File}
\usage{
mosdepth_retrieve_file(
  mosdepth_outputs,
  filetype = c("summary", "global_distributions", "per_base_bed", "per_base_d4",
    "regions_bed", "region_distributions", "quantized_bed", "thresholds_bed")
)
}
\arguments{
\item{mosdepth_outputs}{result of running \code{\link[=mosdepth_collect_output_files]{mosdepth_collect_output_files()}}}

\item{filetype}{type of file to retrieve (string)}
}
\value{
path to chosen filetype (string)
}
\description{
Retrive Mosdepth File
}
