% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdepth_wrapper.R.R
\name{mosdepth_run}
\alias{mosdepth_run}
\title{Mosdepth}
\usage{
mosdepth_run(
  prefix,
  bam,
  threads = NULL,
  flag = NULL,
  fastmode = NULL,
  mapping_quality_threshold = NULL,
  d4 = FALSE
)
}
\arguments{
\item{prefix}{prefix for output files (string)}

\item{bam}{the alignment file for which to calculate depth (string)}

\item{threads}{number of BAM decompression threads. (use 4 or fewer). default: 0.  (string)}

\item{flag}{exclude reads with any of the bits in FLAG set. default: 1796 (string)}

\item{fastmode}{dont look at internal cigar operations or correct mate overlaps (recommended for most use-cases). (string)}

\item{mapping_quality_threshold}{mapping quality threshold. reads with a mapping quality less than this are ignored. default: 0 (string)}

\item{d4}{output per-base depth in d4 format. This is much faster. (boolean)}
}
\description{
simple R wrapper for mosdepth
}
