% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectingMosdepthOutput.R
\name{mosdepth_tabulate_global_distributions}
\alias{mosdepth_tabulate_global_distributions}
\title{Mosdepth Results}
\usage{
mosdepth_tabulate_global_distributions(mosdepth_outputs)
}
\arguments{
\item{mosdepth_outputs}{result of running \code{\link[=mosdepth_collect_output_files]{mosdepth_collect_output_files()}}}
}
\value{
table describing global distributions (what \% of genome is covered by at least coverage > X)
}
\description{
Mosdepth Results
}
