% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectingMosdepthOutput.R
\name{mosdepth_tabulate_per_base_depth}
\alias{mosdepth_tabulate_per_base_depth}
\title{Mosdepth Results}
\usage{
mosdepth_tabulate_per_base_depth(mosdepth_outputs)
}
\arguments{
\item{mosdepth_outputs}{result of running \code{\link[=mosdepth_collect_output_files]{mosdepth_collect_output_files()}}}
}
\value{
table describing per-base depth
}
\description{
Mosdepth Results
}
