% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectingMosdepthOutput.R
\name{mosdepth_tabulate_summary}
\alias{mosdepth_tabulate_summary}
\title{Mosdepth Result Summary}
\usage{
mosdepth_tabulate_summary(mosdepth_outputs)
}
\arguments{
\item{mosdepth_outputs}{result of running \code{\link[=mosdepth_collect_output_files]{mosdepth_collect_output_files()}}}
}
\value{
table summarising mosdepth outputs
}
\description{
Mosdepth Result Summary
}
